chore: split credentials.rs tests + allowlist line-count bumps for agents polish
